third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-018-expected.txt


FAIL CSS Test: @viewport constrained - used width can be zero assert_equals: expected 0 but got 1
